Finished the project!
I am extremely happy to declare this project finally over.
This commit brings several fixes, and now everything should work:

* The function 9, which is used to print out the most frequent value in
the array, now works!
* The functions that were supposed to print out the position of the
maximum and minimum element of the array were not working due to the
fact that the EDX register the logic relied on was not properly
initialized. This has been now fixed.